% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Datos}
\alias{Datos}
\title{Libros de pesca}
\format{
\subsection{\code{Data}}{

Es un dataframe con 20000 observaciones y 40 variables:
\describe{
\item{FT_REF}{Código de marea}
\item{LE_RECT}{Rectángulo estadístico}
\item{ori_al3}{Código fao original de la especie}
...
}
}
}
\source{
Secretaría general de pesca
}
\usage{
Datos
}
\description{
Capturas registradas en El Cachucho durante el año 2020
}
\keyword{datasets}
