% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangulo.R
\name{rectangulo}
\alias{rectangulo}
\title{Función rectangulo}
\usage{
rectangulo(lon1 = -5.8, lon2 = -4, lat1 = 43.3, lat2 = 44.4, b = 0)
}
\arguments{
\item{lon1}{Límite oeste}

\item{lon2}{Límite este}

\item{lat1}{Límite sur}

\item{lat2}{Límite norte}

\item{b}{buffer}
}
\value{
polígono espacial
}
\description{
Crea polígono rectangular y buffer
}
\examples{
# creo un polígono base de El Cachucho de 0.3º de buffer
rectangulo(b=0.3)
# creo un polígono base de Avilés sin buffer
rectangulo(lon1=-6.6,lon2=-5.6,lat1=43.6807,lat2=44.06)

}
