% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_n_ef.R
\name{top_n_ef}
\alias{top_n_ef}
\title{Función top_n_ef}
\usage{
top_n_ef(ef, var_cat = "ori_name_sci", var_num = "LE_KG", n = 10)
}
\arguments{
\item{ef}{eflalo}

\item{var_cat}{nombre de la variable categórica (Especies, artes, años...)}

\item{var_num}{nombre de la variable numérica (Kg, valor económico...)}
}
\value{
df con categorías y los n valores más altos
}
\description{
Recoge de los libros de pesca las n categorías (var_cat) con mayores valores de la variable numérica var_num
}
\examples{
a <- top_n_ef(Datos)
a
}
